backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
python-dotenv==1.0.1
google-genai==1.10.0
google-cloud-aiplatform==1.74.0
google-cloud-storage==2.19.0
google-adk==1.0.0
pillow==11.1.0
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
